Split out resize logic into separate module
